[CI] feat: Build and Publish Multiplatform Snapshots
This commit introduces a new workflow to build and publish multiplatform snapshots of the Gallery application.

The workflow is triggered by pushes to the `dev` branch and approved pull request reviews.

It builds snapshots for the following platforms:

* Android
* Windows
* Linux
* macOS
* iOS
* WasmJs
* Js

For the 'push' event on the `dev` branch, it also publishes a library artifact to GitHub packages.

This workflow improves the development process by automating the build and distribution of snapshots for all supported platforms.