[fluent] Refactor: Use `FluentTheme.colors.borders` for elevation shadow
This commit refactors elevation shadow to use `FluentTheme.colors.borders` instead of hardcoded colors.

It also introduces `strokeShadow` parameter to `elevation` modifier, which allows to customize the shadow color.

For `CircleShape`, the shadow color is set to `FluentTheme.colors.borders.circle`, and for other shapes, it's set to `FluentTheme.colors.borders.control`.

Additionally, the `backgroundSizing` parameter in `Layer` is now optional and defaults to `BackgroundSizing.OuterBorderEdge`.