[fluent] feat: Add a label to the Slider's thumb
This commit adds a label to the Slider's thumb that displays the current value of the Slider. The label is visible when the thumb is being dragged and is positioned above the thumb.

This change improves the usability of the Slider by providing visual feedback to the user about the current value of the Slider.